Renamed route and controller to team
